Update main.cpp
Driver program that takes two arguments from the CLI: digits in dividend, digits in divisor. The program constructs two random numbers with the number of digits specified for each, multiplies them, then runs the product and the random dividend through the div operator. The program then prints the two numbers, the return value of the div operator, the We know the operator is working when the return value, digits in dividend, digits in divisor, and the runtime in milliseconds. If you toy with different inputs you will notice the program runs in linear time for both arguments.